add volumetric_analysis.py
