+++ spateo/plotting/__init__.py
@@ -1,7 +1,5 @@
 """Spatialtemporal modeling of spatial transcriptomics
 """
 
-from .static.three_d_plots import set_mesh, recon_3d
-
 # from .interactive import *
-from .bbs import polygon, delaunay
+from .static import *
