+++ spateo/plotting/static/__init__.py
@@ -1,6 +1,15 @@
 """A complete solution of spatialtemporal dynamics analyses toolkit of single cell spatial transcriptomics
 """
 
-from .three_d_plots import build_three_d_model, easy_three_d_plot, three_d_slicing, compute_volume
-
-from dynamo.plot.space import space
+from .bbs import (
+    delaunay,
+    polygon,
+)
+from .space import space
+from .three_d_plots import (
+    clip_3d_coords,
+    three_d_color,
+    build_three_d_model,
+    three_d_slicing,
+    easy_three_d_plot,
+)
